change chat id to big integer field

+++ database/db/models.py
@@ -63,7 +63,7 @@ class Film(models.Model):
 
 
 class Chat(models.Model):
-    chat_id = models.IntegerField(unique=True)
+    chat_id = models.BigIntegerField(unique=True)
     last_film_id = models.ForeignKey(Film, null=True, on_delete=models.SET_NULL)
     chat_name = models.CharField(max_length=255, default=f"Chat {chat_id}")
     spam = models.BooleanField()
